// src/app/models/user.ts
import { Session } from './session';

declare var $: any;

export class User {
  id: string;
  firstname: string;
  icon: string;
  creation_date: string;
  userName: string;
  title: string;
  created_by_user_id: string;
  enabled: string;
  lastname: string;
  last_connection: string;
  password: string;
  manager_id: string;
  job_title: string;
  last_update_date: string;
  session: Session;

  constructor(jsonData: any) {
    $.extend(this, jsonData);
  }
}